Add CloseEvent to DockItem
This will allow DockItem widgets to perform an action in response to the
titlebar close button being clicked (and ignore this event if desired).
