+++ enaml/qt/qt_dialog.py
@@ -8,7 +8,7 @@
 from atom.api import Typed, atomref
 
 from enaml.widgets.dialog import ProxyDialog
-from enaml.widgets.window import CloseEvent
+from enaml.widgets.close_event import CloseEvent
 
 from .QtCore import Qt
 from .QtWidgets import QDialog
